+++ src/frontends/qtfrontend/tooltabs/projecttab.cpp
@@ -1,5 +1,5 @@
 /******************************************************************************
- *  Copyright (C) 2010-2015 by                                                *
+ *  Copyright (C) 2010-2017 by                                                *
  *    Ralf Lange (ralf.lange@longsoft.de)                                     *
  *                                                                            *
  *  This program is free software; you can redistribute it and/or modify      *
@@ -574,7 +574,7 @@ void ProjectTab::updateMixMode(int) {}
 void ProjectTab::updateMixCount(int) {}
 
 
-void ProjectTab::updateLiveViewFps(int) {}
+void ProjectTab::updateLiveViewFps(double) {}
 
 
 void ProjectTab::updateVideoFps(int) {}
